// src/types/spotify.ts
export interface SinglePlaylistResponse extends PlaylistObjectFull {}
// ...
export interface FollowersObject {
  href: string;
  total: number;
}

/**
 * Image Object
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface ImageObject {
  height?: number;
  url: string;
  width?: number;
}

/**
 * Paging Object wrapper used for retrieving collections from the Spotify API.
 * [](https://developer.spotify.com/web-api/object-model/#paging-object)
 */
export interface PagingObject<T> {
  href: string;
  items: T[];
  limit: number;
  next: string;
  offset: number;
  previous: string;
  total: number;
}
// ...
export interface PlaylistBaseObject {
  collaborative: boolean;
  external_urls: ExternalUrlObject;
  href: string;
  id: string;
  images: ImageObject[];
  name: string;
  owner: UserObjectPublic;
  public: boolean;
  snapshot_id: string;
  type: 'playlist';
  uri: string;
}

/**
 * Playlist Object Full
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface PlaylistObjectFull extends PlaylistBaseObject {
  description: string;
  followers: FollowersObject;
  tracks: PagingObject<PlaylistTrackObject>;
}

/**
 * Playlist Object Simplified
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface PlaylistObjectSimplified extends PlaylistBaseObject {
  tracks: {
    href: string;
    total: number;
  };
}

/**
 * The Track Object in Playlists
 * [](https://developer.spotify.com/web-api/object-model/)
 */
export interface PlaylistTrackObject {
  added_at: string;
  added_by: UserObjectPublic;
  is_local: boolean;
  track: TrackObjectFull;
}